refactor: rename runners -> root_runner
Preparation for #168
